+++ ElvUI_MerathilisUI/Core/Changelog/6.74.lua
@@ -1,8 +1,17 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[674] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	RELEASE_DATE = "28.09.2025",
+	FIXES = {
+		"[ElvUI]: Required Version >14.00",
+		"[MicroBar]: Fix random hearthstone on the MicroBar",
+		"[Who Frame]: Fix a lua error on the Who Frame",
+	},
+	NEW = {
+		"[AIO skin]: Added new skin for Advanced Interface Options",
+	},
+	IMPROVEMENTS = {
+		"[Chat]: Rewrote the Emote Module completly",
+		"[Skins]: Updated a lot of skins with some updated functions",
+	},
 }